// src/travels/enums/travels.enum.ts
export enum CompanyName {
  MAHAN = 'mahan',
  SIROSAFAR = 'sirosafar',
  MEHR = 'mehr',
}

export enum Towns {
  MASHHAD = 'mashhad',
  TEHRAN = 'tehran',
}

export enum ServiceTypes {
  PLAIN = 'plain',
  BUS = 'bus',
  CAR = 'car',
}

export enum Level {
  CLASIC = 'clasic',
  VIP = 'vip',
  VIP2 = 'vip2',
}